Create all_mp4_wSkip_placeholder
This uses whisper-faster on Windows without a GPU to transcribe video files. It checks if a transcription has already been done and if none is found it starts it. This also allows for multiple instances of the powershell script to be run on the same server by placing a placeholder file while the transcription is taking place of one file, hence skipping over that file and starting on the next file.